Reverted to the former version of the CAENVME library (usable within a SLC4/RHEL 4 environment) ; Matched the CMakeLists.txt configuration file to the LIP PC configuration

+++ src/VME_BridgeVx718.cpp
@@ -78,7 +78,7 @@ namespace VME
   void
   BridgeVx718::InputRead(CVInputSelect input)
   {
-    unsigned int data;
+    short unsigned int data;
     if (CAENVME_ReadRegister(fHandle, cvInputReg, &data)!=cvSuccess) {
       std::ostringstream o;
       o << "Failed to read data input register #" << static_cast<int>(input);
